# api/swagger/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateStreamRequest": {
        "properties": {
          "max_age": {
            "type": "string"
          },
          "max_bytes": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "replicas": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "retention": {
            "enum": [
              "limits",
              "interest",
              "workqueue"
            ],
            "type": "string"
          },
          "storage": {
            "enum": [
              "file",
              "memory"
            ],
            "type": "string"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "name",
          "storage",
          "subjects"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "PurgeStreamRequest": {
        "properties": {
          "sequence": {
            "type": "integer"
          },
          "subject": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "PurgeStreamResponse": {
        "properties": {
          "message": {
            "type": "string"
          },
          "remaining": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "StreamConfigResponse": {
        "properties": {
          "max_age": {
            "type": "string"
          },
          "max_bytes": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "replicas": {
            "type": "integer"
          },
          "retention": {
            "type": "string"
          },
          "storage": {
            "type": "string"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "StreamResponse": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/StreamConfigResponse"
          },
          "state": {
            "$ref": "#/components/schemas/StreamStateResponse"
          }
        },
        "type": "object"
      },
      "StreamStateResponse": {
        "properties": {
          "bytes": {
            "type": "integer"
          },
          "consumers": {
            "type": "integer"
          },
          "delete_marks": {
            "type": "integer"
          },
          "first_seq": {
            "type": "integer"
          },
          "first_ts": {
            "type": "string"
          },
          "last_seq": {
            "type": "integer"
          },
          "last_ts": {
            "type": "string"
          },
          "messages": {
            "type": "integer"
          },
          "num_pending": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "UpdateStreamRequest": {
        "properties": {
          "max_age": {
            "type": "string"
          },
          "max_bytes": {
            "type": "integer"
          },
          "replicas": {
            "maximum": 5,
            "minimum": 1,
            "type": "integer"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "NATS Horizon Project",
      "url": "http://github.com/nats-horizon"
    },
    "description": "Comprehensive API for monitoring and managing NATS JetStream instances.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://opensource.org/licenses/Apache-2.0"
    },
    "termsOfService": "http://swagger.io/terms/",
    "title": "NATS Horizon API",
    "version": "1.0"
  },
...
  "openapi": "3.1.0",
  "paths": {
...
    "/streams": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StreamResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List all streams",
        "tags": [
          "streams"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStreamRequest"
              }
            }
          },
          "description": "Stream creation request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Create a new stream",
        "tags": [
          "streams"
        ]
      }
    },
    "/streams/{name}": {
      "delete": {
        "parameters": [
          {
            "description": "Stream name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Delete a stream",
        "tags": [
          "streams"
        ]
      },
      "get": {
        "parameters": [
          {
            "description": "Stream name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get stream details",
        "tags": [
          "streams"
        ]
      },
      "put": {
        "parameters": [
          {
            "description": "Stream name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStreamRequest"
              }
            }
          },
          "description": "Stream update request",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Update a stream",
        "tags": [
          "streams"
        ]
      }
    },
...
    "/streams/{name}/purge": {
      "post": {
        "parameters": [
          {
            "description": "Stream name",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurgeStreamRequest"
              }
            }
          },
          "description": "Purge options",
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeStreamResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Purge messages from a stream",
        "tags": [
          "streams"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}